+++ src/index.ts
@@ -1,3 +1,3 @@
-export * from "./client";
+export * from "./WABA_client";
 export * from "./types";
 export * from "./webhooks/client";
